third_party/blink/web_tests/external/wpt/html/semantics/interactive-elements/the-dialog-element/dialog-focusing-steps-disconnected-expected.txt

This is a testharness.js-based test.
FAIL dialog.show(): focusing steps should not change focus on disconnected <dialog> assert_equals: Focusing steps should not change focus expected Element node <input></input> but got Element node <body>
<input>
<script>
test(function() {
  const outerIn...
PASS dialog.showModal() should throw on disconnected <dialog>
Harness: the test ran to completion.
